Moving internal raw API response models to models.internal sub-package
Separate private[yfinance4s] raw response types (YFinanceQueryResult,
YFinanceQuoteResult, etc.) from public domain models by relocating them
to a dedicated models.internal package, making the public/internal API
boundary explicit in the package structure.
