+++ core/shared/src/main/scala/org/coinductive/yfinance4s/Analysts.scala
@@ -5,6 +5,7 @@ import cats.syntax.flatMap.*
 import cats.syntax.functor.*
 import org.coinductive.yfinance4s.Mapping.*
 import org.coinductive.yfinance4s.models.*
+import org.coinductive.yfinance4s.models.internal.*
 
 /** Algebra for analyst data (price targets, recommendations, estimates, etc.). */
 trait Analysts[F[_]] {
