+++ core/shared/src/main/scala/org/coinductive/yfinance4s/Holders.scala
@@ -5,6 +5,7 @@ import cats.syntax.flatMap.*
 import cats.syntax.functor.*
 import org.coinductive.yfinance4s.Mapping.*
 import org.coinductive.yfinance4s.models.*
+import org.coinductive.yfinance4s.models.internal.*
 
 /** Algebra for holders and insider data. */
 trait Holders[F[_]] {
