+++ core/shared/src/main/scala/org/coinductive/yfinance4s/models/DividendEvent.scala
@@ -1,6 +1,6 @@
 package org.coinductive.yfinance4s.models
 
-import org.coinductive.yfinance4s.models.YFinanceQueryResult.DividendEventRaw
+import org.coinductive.yfinance4s.models.internal.YFinanceQueryResult.DividendEventRaw
 
 import java.time.{Instant, ZoneOffset, ZonedDateTime}
