+++ core/shared/src/main/scala/org/coinductive/yfinance4s/models/SplitEvent.scala
@@ -1,6 +1,6 @@
 package org.coinductive.yfinance4s.models
 
-import org.coinductive.yfinance4s.models.YFinanceQueryResult.SplitEventRaw
+import org.coinductive.yfinance4s.models.internal.YFinanceQueryResult.SplitEventRaw
 
 import java.time.{Instant, ZoneOffset, ZonedDateTime}
