+++ core/shared/src/main/scala/org/coinductive/yfinance4s/models/internal/YFinanceHoldersResult.scala
@@ -1,8 +1,8 @@
-package org.coinductive.yfinance4s.models
+package org.coinductive.yfinance4s.models.internal
 
 import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
-import org.coinductive.yfinance4s.models.YFinanceQuoteResult.Value
+import org.coinductive.yfinance4s.models.internal.YFinanceQuoteResult.Value
 
 /** Raw API response for holders data from Quote Summary endpoint. */
 private[yfinance4s] final case class YFinanceHoldersResult(
